Throw configuration errors instead of continuing
In my previous implementation, I was checking for some non-sensical
configurations in the GlobalChi2Fitter and adjusting them to make sense.
However, this is apparently not recommended as it can mess with
configuration in the trigger, and the recommended way of going about it
is to simply throw a configuration error. This commit does exactly that.