Remove calls to old covariance matrix getter
As it stands, most of the external interactions with the covariance
matrix of the pattern track parameters happens through another C-style
array getter. This approach is not compatible with the Eigen-based
solution we're moving to, which is also used in the ParametersBase
class. Therefore, we first migrate all uses of this old getter to the
other getter which returns an Eigen object. Note that this commit is a
very likely performance regression, since it involves a lot of copying.
This will be resolved later.