Replace covariance matrix by ptr to Eigen matrix
The existing implementation of the covariance matrix data member keeps
an array of 15 elements to represent the upper (or lower) triangular
part of the matrix. This is not compatible with the ParametersBase
class, which uses an Eigen object instead. Specifically it uses an AMG
symmetric matrix. This commit replaces the data member in the pattern
track parameters with a similar object.