Slightly hacky fix for failure to find CaloExtension
This fix gets around the ongoing problem with rare crashes when determining muon etcone isolation due to failing to find a CaloExtension if the dynamic ID alignment is enabled by adding a second instance of the ParticleCaloExtensionTool to the MuonCreatorTool. Now there is one instance for combined (and MS) tracks, and one for ID tracks: the former starts extrapolating from the perigee, the latter from the ID exit. The CaloIsolationTool no longer performs a separate extrapolation, but uses the result of the one in the MuonCreatorTool.  Note that should the CaloExtension not be found, there will still be a crash in the CaloIsolationTool, but tailoring the starting point correctly in the MuonCreatorTool should ensure that the CaloExtension is found.  The real solution to the crash is to fix the Extrapolator, as discussed in ATLASRECTS-5674, but this will hopefully avoid the problem for the moment.
The real source of the crash is