tauRec: update of ESD and AOD tau containers
Hello,

This MR is updating the list of ESD and AOD tau containers.
Some containers have been renamed for the sake of harmonisation and clarity: TauPi0SubtractedClusters becomes TauInitialPi0Clusters (there is no subtraction performed here), and finalTauPi0s becomes TauFinalPi0s.
The containers are now all added via TauESDList and TauAODList, and the use of objKeyStore has been discontinued. It is unclear why objKeyStore was still in use, it is confusing that some containers are stored via e.g. "objKeyStore.addManyTypesStreamESD", while some others are saved via "TauESDList".
Note: a final renaming might occur very soon, because we no longer need to define a "TauPi0Clusters" container (which is a subset of TauPi0SubtractedClusters = TauInitialPi0Clusters, that we used to store on disk), now that we store TauPi0SubtractedClusters = TauInitialPi0Clusters (to become simply "TauPi0Clusters").
To be done in a follow-up MR.

Cheers,
Bertrand