Remove InDetCandidate selection for muon triggers
Updating the muon InDetCandidateAlg to be able to skip using the InDetTrack selection tool, and updating the trigger configuration to not use apply the InDetTrack selection. It turns out that this cuts out quite a number of ID tracks, and does seem to have some impact on low pT muons.
Also added as mall change to initialise the ReadHandle for teh forward track particles only if it's actually needed.