tauRec: TauCommonPi0Cells thinning (first version)
Hello,

As discussed in ATLASRECTS-5684 and ATLTAU-1733, this MR is implementing the easiest mechanism we have to decrease the size of the tau pi0 cell container in the xAOD.
We simply keep the cells that belong to pi0 topoclusters.
It is not the optimal solution, though, but merging TauCommonPi0Cells and AllCalo will require more work.
This is under study.

Cheers,
Bertrand