Cache position and momentum of pattern parameters
In the current implementation of the pattern track parameters methods
for the position and momentum of a set of parameters, the resulting
vectors are created transiently. That is to say, they are not stored
anywhere and are created on-demand when the accessor methods are called.
This presents two problems in regards to the unification with the
ParametersBase class: Firstly, it means that the pattern parameters do
not have the requisite data members that the base parameters have.
Secondly, it means that the signature for the getter methods is
different, because the pattern parameters return an object directly
while the base parameters return a reference.  This commit, therefore,
adds some cache values that contain the position and momentum, such that
they can be accessed by reference.