Embed charge object in pattern track parameters
Currently, the pattern track parameters return the value for charge
using the value of the q/p. This works fine, but it does not align with
the implementation of the base parameters which use a Trk::Charged (or
Trk::Neutral) object. This commit redoes the internals of the pattern
track parameters to have such an object embedded in it.