Use smart pointers for pattern parameter surface
Currently, the pattern track parameters contain a normal non-owning
pointer to a surface. This works fine, but again, it doesn't align with
what the base parameters use. The base parameters use a unique pointer
with a non-standard destructur, encapsulated in the SurfaceUniquePtrT
class. This commit ensures that the pattern track parameters implements
its surface in the same way.