Update position after momentum, not before
In the existing implementation, the `ParametersT` class, when updating
the track parameters, will update the position first and the momentum
second. This is a potentially problematic implementation, however, as
there are surface types for which the position depends on the momentum.
For such surface types, the current implementation will update the
position according to the old and potentially outdated momentum, instead
of the new momentum. This commit resolves this issue by reordering these
computations such that the momentum is updated before the position.