Force position update for appropriate surfaces
Currently, it is possible for the `ParametersT` class to update the
momentum of the track parameters without updating the position. However,
for some surface types the computation of the position depends on the
momentum. This means that it is possible for the position to go out of
sync with the momentum, leading to incorrect results. This situation is
most likely quite rare.