Trigger matching in DAOD_PHYS in master, part 1
This is the first step in getting trigger matching to work in DAOD_PHYS in master. It restores the machinery for getting the required list of trigger that was broken by changes to the HLT menu metadata in master. It does not yet restore the trigger matching itself, which will need further work. Consequently this MR has no impact on the output of DAOD_PHYS, because the active lines are still commented.