Trigger matching in DAOD_PHYS in master, part 2
This MR makes a small fix to the TriggerMatchingHelper such that it is compatible with Python 3. It restores the remaining lines in DAOD_PHYS needed to restore the trigger matching decorations to DAOD_PHYS.

With this change the trigger matching information as produced by DAOD_PHYS in release 22 should now be equivalent to release 21.