AthContainersRoot: Don't use GetClassSharedLibs.
getClassIfDictionaryExists() was calling GetClassSharedLibs() to try to tell
if a class was present in a rootmap file, and thus avoid triggering
autoparsing.  However, in recent root versions, GetClassSharedLibs
itself can trigger autoparsing, followed by an unload which can
confuse things.  GetClassSharedLibs is now doing more harm than good,
so remote the call to it.

Still need to see if we can find a reliable way of failing
a TClass lookup rather than triggering autoparsing.

See ATEAM-662.