AthenaMT condition data handling: allow overriding IOVs of individual folders
IOVDbSvs patched such that if IOV for a given folder is overriden,
then it sets infinite validity for the condition object fetched
from the database. As a result, the folder cache never gets updated, and the
object fetched from the database remains valid for the duration of the job.

Addresses ATEAM-649