Make Muon Segment Finding Algs Reentrant
Migrating the muons egment finding algorithms to be reentrant (ATR-22181). Mostly changing AthAlgorithm to AtReentrantAlgorithm, making relevent functions const in the algorithms and associated tools, and passing the event context to the data handles. (Note the event context should also be passed down through the tools to the data handles included in tools. Will do this in separate merge requests, as there is quite a labyrinth of tools to navigate and this MR is already starting to get a bit large).