Make changes to ParameterBase member qualifiers
This commit makes the following two qualifier changes to the
ParametersBase class:

1. It makes the position, momentum, and charge getter methods virtual.
This may introduce some very minor performance degrading effects due to
extra vtable lookups, but the overall effect of this should be extremely
small.

2. It makes the data members for the position, momentum, and charge
definition mutable. This is to enable future caching, where these
members may need to be modified in const getter methods.

Both these changes are made to aid the unification of the pattern track
parameters with the base parameters. The pattern parameters have a
fundamentally different lazy evaluation approach to momentum, position,
and charge.