MuonHoughPatternTools: Remove useless code.
Remove a useless call to stationCode; the result was being overwritten
by the next statement.

That call was also giving an assertion failure in dbg builds because
we were using a TGC identifier with the MDT ID helper; the eta field
of some TGC identifiers is out of range for MDT.

There is another place in this file where call stationCode with an RPC
identifier.  This in principle has the same problem.  However, i have not
seen this tripping an assertion, so leave this for now.