TileConditions+TileMonitoring: Modify TMDB calbiration tool interface
Interface of TMDB calbiration tool has been modified to use raw channels
instead of enums, because mapping from channels to TMDB cells is different
in Tile LB for odd/even modules and using these enums in this case is misleading.
Thresholds enums have been removed also since they are different for run 2 and run3
and in this case they are misleading also.

This tool seems to be used only in one test and monitoring tool,
so they have been modified accordingly.