tauRec: improved tau cell thinning
Hello,

This MR is implementing an improved version of the tau pi0 cell thinning in xAODs (ATLASRECTS-5684, ATLTAU-1733).
The pi0 clusters are initially built from cells stored in a separate container (TauCommonPi0Cells).
An algorithm (ClusterCellRelinkAlg) is introduced to recreate pi0 clusters with cells from the main cell container (AllCalo), thereby eliminating the overhead from our separate cell container.
The ESD and AOD contents have been adjusted accordingly: TauCommonPi0Cells is dropped, and the relinked pi0s are written out instead of the initial pi0s.
There will be a follow-up MR that will drop some other pi0 container we used to use in TauPi0ClusterCreator, which should no longer be needed.
For now, only the minimum amount of TauPi0ClusterCreator code was changed.

Another algorithm (TauCellThinningAlg) is introduced to save the cells from the tau seed-jet topoclusters in xAODs.
Only clusters within 0.2 of the tau are considered, to limit the xAOD size increase.
Note: in the xAODs, we are still missing CaloCalTopoClusters_links (the CaloClusterCellLinkContainer) for the tau seed-jet clusters.
An attempt at thinning this container was made, but it was not successful.

Cheers,
Bertrand