temporary workaround for xAOD dictionary issues in AnalysisBase
This is meant to be just a temporary workaround, as currently there
are issues with loading the xAOD dictionaries in AnalysisBase that can
be very hard for users to debug and understand.  And particularly
since a more permanent fix is currently under development it seemed
easiest to import the corresponding workaround from 21.2 for the time
being instead of integrating all the workarounds into each of the
current unit tests and then taking them out again in a few weeks.

This is mostly the same code as in LoadDictionaries() in 21.2, but I
replaced the use of a preprocessor macro with a simple for loop, and
slightly adjusted the messaging to match the rest of the function.