tauRec: drop unused pi0 container
Hello,

This MR is hopefully the last in the tau pi0 MR series :-)
Following the recent reorganisation of our pi0 containers, we are dropping a container that is no longer used.
We are now left with only 2 pi0 containers in the tau reconstruction: TauInitialPi0Clusters (transient, which results from the topoclustering of cells from the TauCommonPi0Cells container), and TauPi0Clusters (persistent, which are identical except that their cells point to the AllCalo container).
This MR does not change the reconstruction output.

Cheers,
Bertrand