AthExStoreGateExample: Fix default name of handle keys
With the change to the `DataHandle` type for handle keys, setting the
default key name in the constructor no longer works. This is because
during `genconf` generation the instance name is set to "DefaultName"
and unlike for basic types, this value will be set in the job options
catalogue even if it is the default.