tauRec: add flag for maximum allowed number of tau tracks in reconstruction (skimming)
Hello,

This MR is introducing a flag to skip tau candidates that have too many tau tracks.
The number of tracks corresponds to the number of tracks classified as good tau tracks by the tau track classifier.
Currently, we already run part of the tau reconstruction (tau substructure reconstruction) for candidates that have 1-5 tracks.
The new flag can be used to run the whole reconstruction only for taus with e.g. 0-5 tracks.
We would save some CPU and disk (relevant ticket: ATLASRECTS-5684).
The main motivation is that taus with >3 tracks will never be used in analyses.
0-track candidates should be kept as they are good candidates for e.g. neutral LLP->tau. The 1-3 track candidates are for physics. Keeping up to e.g. 5-track candidates gives some flexibility to retune and improve the track classifier at xAOD level, although that is somewhat unlikely to happen.

Cheers,
Bertrand