LArByteStream: Further thread-safety fixes.
Hid2RESrcID depends on the old, thread-unsafe LArCablingLegacyService.

We can't get rid of it completely, because it's required by interfaces
still used by TrigT2CaloCommon.

Refactor so that it won't affect the thread-safety of the rest of this package.

Split out the part of Hid2RESrcID required by FullEventAssembler into
a separate base class RodRobIdMap.

Split the remaining interfaces of Hid2RESrcID into thread-safe and
thread-unsafe versions.  Only the thread-safe functionality will be needed
from this package; the thread-unsafe versions are needed only
by TrigT2CaloCommon.