checkxAOD: move "Stau" collections from tau to muon domain
Hello,

This MR is simply migrating the "ExtrapolatedStauTrackParticles" and "CombinedStauTrackParticles" containers from the tau category to the muon category.
"Staus" is currently under muons, so I assume that's a fair thing to do, these collections are not produced by tau reconstruction.
Unfortunately these are small collections, so we do not transfer much disk footprint :-)

Cheers,
Bertrand