tauRec: thinning of CaloCalTopoClusters_links for tau clusters
Hello,

This MR adds the functionality to thin the CaloCalTopoClusters_links container and only keep the entries associated with tau clusters.
This container is needed for the tau cluster-to-cell navigation in xAODs, but it too large to be written out without thinning.
As discussed in ATLTAU-1733, the CaloClusterCellLinkContainer thinning was not supported so far, and Scott just implemented it (not merged yet).
Once it's in Athena, the CaloCalTopoClusters_links size should be reduced significantly (I'm already adding the container to the xAOD output to minimise the number of MRs, so we expect a jump in xAOD size, ATLASRECTS-5684).
According to q221 run on 40 events, thinning reduces the container size by a factor of 2.5 (3.0 kb/evt after thinning).

Cheers,
Bertrand