Fix for RPC ROB based decoding
Passing the vector of hashIDs to decode to the processPad function so that we don't decode the full RPC detector when it isn't needed (this was already done correctly for the PRD based decoding, but for the ROB based decoding the vector being passed was always empty). Because we no longer decode the full RPC detector, this changes slightly the muons we reconstruct in the trigger (and the trigger counts) so the references have been updated accordingly.