Thread safe handling of CORAL transactions by RDBAccessSvc
Given CORAL transaction context seems to be global, one thread can
commit a transaction, which is active at that moment in another thread.
Worked around this issue by starting readonly transaction when a
session is created, and committing that transaction only when the
corresponding session is closed.

Addresses the issue discussed in ATLASRECTS-5737.

Also removed several redundant include statements.