tauRec: phase out tau electron veto BDT (ATLTAU-1719)
Hello,

This MR is dropping a deprecated tau electron veto algorithm based on a BDT classifier. It is superseded by the RNN eVeto.
We will be phasing out a few more tools soon in separate MRs, such as the BDT tau ID superseded by the RNN ID since long.
The motivation is mostly to see if anything breaks in e.g. monitoring tools.

The only changes to the xAOD output are, as expected:
```
Py:diff-root      WARNING the following variables exist only in the old file !
Py:diff-root      WARNING  - [TauJetsAuxDyn.BDTEleScoreSigTrans]
and e.g.
000.TauJetsAux.isTauFlags.0 29360129 -> 1 => diff= [49.99999659%]
000.TauJetsAux.BDTEleScore.0 0.9994122385978699 -> 0.0 => diff= [50.00000000%]
```

The size reduction of the TauJets collection is negligible (3.021->2.969 kb/evt in MC ESD->xAOD test).

Cheers,
Bertrand