MGC Beta Feature: Reproducer script
This implements a new feature in MadGraphControl.

We now try to track commands sent to the shell as they are stacked
during MadGraphControl processing. A few helpers have been added for
this tracking (e.g. stack_process). If the job crashes and an error is
going to be raised, then we try to write out a script that will locally
reproduce the error.

If this works well, it should be *much* easier for us to help people
with debugging offline.