tauRec: drop TauEleOLRDecorator tool
Hello,

This MR is dropping the TauEleOLRDecorator tool from the tau reconstruction (ATLTAU-1719).
This tool was used for the old electron-faking-tau veto that relied on the likelihood of the reconstructed electron matched with the tau.
We now have an RNN eVeto solely based on variables from the reconstructed tau, so we no longer need to match electrons with taus.
This is good because we are dropping the dependency of tau reconstruction on electron reconstruction introduced by the SG::ReadHandle<xAOD::ElectronContainer>.
Still, in the upcoming R22 RNN eVeto tuning, we will study if there is any gain from a (electron GSF track pt)/(tau track pt) variable.
Only in case we find significant improvement in the electron rejection, we might reintroduce some form of tau-GSF track matching matching in the RNN code.

As expected, the change in the xAOD reconstruction output is:
```
Py:diff-root      WARNING the following variables exist only in the old file !
Py:diff-root      WARNING  - [TauJetsAuxDyn.electronLink.m_persKey]
Py:diff-root      WARNING  - [TauJetsAuxDyn.electronLink]
Py:diff-root      WARNING  - [TauJetsAuxDyn.electronLink.m_persIndex]
```
and e.g.
```
001.TauJetsAux.EleMatchLikelihoodScore.1 -1.8590551614761353 -> 0.0 => diff= [50.00000000%]
002.TauJetsAux.isTauFlags.1 4030464001 -> 4030464000 => diff= [0.00000001%]
```

Cheers,
Bertrand