CxxUtils: Small improvements for CachedValue::isValid().
CachedValue::isValid() should be optimized for the case where the value
is already valid.  In that case, we were doing two comparisons, which
likely resulted in forward branches.  Adjust so that we need only
make one comparison in this case, and use ATH_LIKELY to make it
less likely that we need to branch in this case.