Fixing warnings in derivation jobs
This fixes a handful of warnings:

-) From EventInfoCnvTool, when certain truth collections are not
available. This is expected in any job running on EVNT

-) In PrimmaryDPDFlags, some lines were duplicated (completely
identical), leading to a warning when the flags are imported

-) In RecExCommon_flags, athena common flags values were checked
independent of whether they had statusOn. Now the code checks for
statusOn before checking the value of the flag

-) In Reco_tf.py, both digitization and overlay were loading the basic
digitization flags into the transform. Now only the overlay portion does
the loading of these flags.