tauRec: fix flag check
Hello,

There was a recent jump in CPU for the first tau algorithm we run, while there was no change in any tool:
https://atlaspmb.web.cern.ch/atlaspmb/arch-mon-fullchain_mc16_ttbar_valid_13tev_25ns_mu40/?test=cpu/master__x86_64-centos7-gcc8-opt#RAWtoESD/tau

I suspect it's due to
https://gitlab.cern.ch/atlas/athena/-/merge_requests/37904/diffs#491ed5996e49b45b60c269a6852c901e7f54cf9c_1026_1022
where rec.doESD should rather be rec.doESD().
As the check fails, we might not be using the cached track-calo extrapolation, which may cause the CPU increase.
I'm running some test to verify this. Results in a few minutes.

Cheers,
Bertrand