tauRecTools: associate cluster in CaloCalTopoClusterContainer to the tau candidate
For Topo seed jets, the cluster links in the tau candidate point to the
clusters in a shallow copy container, which will not be available in AOD
of R22. To fix this, the cluster links point to the cluster in the
original container.

For PFlow seed jets, the cluster links in the tau candidate point to the
PFOs(neutral or charged). It is confusing. Now the cluster links point
to the corresponding cluter in the PFO object.

Previously, only clusters within 0.2 cone of tau candidate is stored. In
the tau reconstruction, we need all the clusters. Thus the cut of dR is now
removed.

The clusters could have energy <=0, which will be skimmed in AOD. The
link then becomes invalid. Since we do not want an invalid link, all
the clusters are requied to have positive energy.