Return AMG vectors by value in Tracking
There are some cases in which the references returned by the track
parameter position and momentum methods are returned by reference for a
second time. Given the fact that we're attempting to move the
aforementioned methods to return by value, this means that these return
by reference methods would become memory unsafe. We change these methods
to be return by value themselves to resolve this issue.

This commit in particular resolves the issues in the Tracking realm.