RootUtils: cmake fix for clang compilation
Only the RootUtilsPyROOT library should depend on the pyroot libraries
(ROOTTPython and cppyy).  In particular, the RootUtils library should not
depend on these libraries.

Things were failing in the clang build because RootUtils was linking
against cppyy but _not_ against python, so things downstream were failing
to find python symbols.

(In principle this is wrong for the gcc build as well; not entirely
sure why it was working there.)