merge xAOD::ReturnCode into StatusCode, migrate all users
This has been a long time coming, but one of the reasons for having
AsgMessaging as a separate package from AsgTools is to allow
xAODRootAccess to depend on it and use the same status codes and
messaging as the rest of analysis base.  This should take part of the
former.

I had to add the RECOVERABLE code to the main status code, but since
that is also part of the Gaudi StatusCode that seems like a reasonable
change/extension.

Apart from that this is mostly a global search replace to move all
users from TReturnCode to StatusCode, but apart from fixing up a
couple of places related to the status code itself I did not try to
fix/improve any code in the process.

Note that this also revealed a number of tests that failed because
they used RETURN_CHECK() incorrectly.  This commit doesn't try to
address that.