Enabling truth jets in TRUTH derivations
This gets truth jet building working again in truth derivations. The
state after this MR is:

-) The correct large-R jets are built and have the required decorations
-) The correst small-R jets are build, but they currently lack the
desired truth flavor decorations
-) As discussed elsewhere, some extra branches exist in truth DAODs made
with master, because branch dropping seems to not be working quite right
yet
-) There are errors coming from CutFlowSvc to do with a mis-match in the
number of weights. I'm not quite sure how to solve this one; it looks
like CutFlowSvc has moved and been re-worked in master, so I might ask
for a little help with this one.

Note that I've disabled most of the jet-related items for derivations
running off of AODs. That's because the truth jet building code
currently conflicts with what the jet group is running, and it's not yet
clear to me how to fix this all. I hope the jet group will clean up a
bit in the next few days/weeks, and then we'll be able to harmonize.