AthContainers: Fix variable selection.
In copyAuxStoreThinned, we can veto variables copied to the output
based in ThinningInfo.  But this should only be done for dynamic variables,
as we always want the full static store to be written.

We were doing this by testing whether a variable was dynamic in the source
store.  However, this fails in the case where the source store is a static
store object and the destination store is a generic AuxContainerBase object.
(This typically happens during derivation when we want to write
all variables as dynamic.)

Fix by also considering a variable to be dynamic if it doesn't exist
in the destination store before we do the copy.