AthenaServices: Fix variable selection.
AthenaOutputStream writes into ThinningInfo the set of variables to be
vetoed on output.  Only dynamic variables should be selected/vetoed
this way, so we were only actually selecting dynamic variables, marking
the rest as vetoed.  The code in copyAuxStoreThinned would then see
that variables were not dynamic and ignore the veto information for them.

However, that turns out to be incorrect for the case where we write
all variables as dynamic (as used during derivation).  So here properly
apply the selection rules for all variables, even static ones.