tauRecTools: retrieve the orignal cluster only  when jet performs vertex correction (ATLASRECTS-5763)
In TauClusterFinder, we retrieve the orignal cluster via a decorated
link. However, if there is no vertex correction in jet reconstruction, then the
returned original cluster will be nullptr, which result in a crash.

To fix this, we retrieve the orignal cluster only when jet performs
vertex correction.

In addition, a flag of SkipNegativeEnergy is added to control the
removal of cluster with rawE <= 0