Add initial version of G4AtlasActionInitialization
`G4AtlasActionInitialization` derives from `G4VUserActionInitialization`.
Classes based on  `G4VUserActionInitialization` are the recommended way
of initializing `G4UserActions` in Geant4MT. Switching to this approach
should make our Athena interface to Geant4 more standard. Currently the
class just wraps calls to the `UserActionSvc`. Ideally `G4RunActions`
would be initialized for the master thread also via the `BuildForMaster`
call. This functionality was missing in the current implementation, so
has not been added here in order to reproduce exactly the current behaviour.
Adding it requires significant refactoring which is best handled in a
separate merge request.