IOVSvc: Fix bug in incident listener
The detection of the first call to the incident listener was not working
correctly. The code relied on the value of `m_first`, which was being
modified during the execution of the function. So code executed after
toggling `m_first` would see the wrong value. Use the local variable
`first` consistently.