Link TrkExTools and Utils with pattern parameters
The track extrapolation packages TrkExTools and TrkExUtils are related
to the pattern track parameters, but neither of these packages are
properly linked against the pattern track parameter packages. This
commit ensures that both of these packages are linked publicly against
the pattern parameters. This was not giving problems before - I'm not
exactly sure why, but I ran into linking problems with some changes to
the pattern code.