xAODRootAccess: Don't crash opening a schema-evolved file.
Don't crash opening a file that uses some kinds of automatic schema evolution.
If the type of a member changes from one class to another
(for example, from std::vector<double> to std::vector<float>),
then GetExpectedType will give us the on-disk class.
To get the class we use in memory, we need to use GetCurrentClass.