tauRec: enable RNN track classification by default
Hello,

This MR is switching on the RNN track classification by default in the R22 tau reconstruction.
So far, we were using the R21-style set of 3 BDTs for tau track classification, but this won't be used in R22.
We were keeping it for the sake of having stable tau distributions in PHYSVAL and DQ checks.
Now that the 22.0.20 release is built for the November (final?) PHYSVAL check, we can switch this on.
It will be interesting to check the impact on SPOT plots (ATLASRECTS-5684), as many tau containers are only built for taus with 1-5 tracks (where the track multiplicity is that after track classification).

Cheers,
Bertrand